Add files via upload
This project is a fun "Truth or Dare" game where you can insert your name and choose between telling the truth or doing some kind of dare. Depending on your answer, it may be a special question you must answer or a challenge for you to complete. The game will keep count of how many truths and dares you have done and will be able to go on playing as long as you wish. It is conceived to be a light form of fun when one wants to have a few laughs with friends.