Add files via upload
The "Personality Palette" game is an interactive and playful program where you explore a creative analysis of your personality based on your favorite season, color, and animal. After entering your name and deciding to play, you begin by selecting your favorite season. This choice, along with your favorite color from a wide array of options, and an animal from a list including an elephant, lion, dolphin, and more, leads to a fun and imaginative personality prediction. The game generates a unique description reflecting the characteristics of your chosen season, color, and animal. Whether you decide to play or not, the game concludes with a warm goodbye, wishing you a nice day. It's a lighthearted way to explore different aspects of your personality or simply enjoy some creative fun with friends.