Add files via upload
Welcome to the Number Funhouse! where the sky is filled with stars waiting to be counted! This simple yet fun program randomly generates a sky full of stars (up to 100) and challenges you to guess how many are twinkling above. You’ll see the stars displayed, and it’s up to your sharp eyes to give the correct number. Guess right, and you’ll be the star-counting champion! Guess wrong, and the game will let you know how many stars there really were, encouraging you to try it again.